feat: extract cells to 2 (3) cells
